[Example] Initial commit of a bazel platform / toolchain transition example with generated source code (#1)
In this example, a small binary is created to understand how transition in combination with generated source code files act in bazel. The outcome of this example is the following:

with `bazel build //...` which build on host1 platform, the generated files behind the transition (host2) are only genereated if a library depends on it.
if you remove the dependencies in BUILD on the returner library, and fix main.cpp, the return_....hpp are not generated by bazel build //... for the host2 toolchain.
If you run an bazel aspect over the complete tree, then they are genrated, but only one by one.

It belongs to internal ticket FSD-180.